Cool

+++ src/core/Input.cpp
@@ -28,7 +28,7 @@ bool Input::KeyDown(const char* key)
 {	
 	bool answer;
 
-	for (int i = 0; i < sizeof(keys) / sizeof(const char*); i++)
+	for (int i = 0; i < (int)(sizeof(keys) / sizeof(const char*)); i++)
 	{
 		if (keys[i] == key)
 			answer = keyBools[i];
